Extracting logic from controller to a service
