+++ lib/Onurb/Bundle/YumlBundle/Controller/YumlController.php
@@ -22,68 +22,10 @@ use Onurb\Bundle\YumlBundle\Curl\Curl;
  */
 class YumlController extends Controller
 {
-    const YUML_POST_URL = 'http://yuml.me/diagram/plain/class';
-    const YUML_REDIRECT_URL = 'http://yuml.me/';
-
     public function indexAction()
     {
-        return $this->redirect($this->getGraphUrl($this->makeDslText()));
-    }
+        $yumlClient = $this->get('onurb.doctrine_yuml.client');
 
-    /**
-     * @return string
-     */
-    private function makeDslText()
-    {
-        return $this->generateGraph($this->getClasses());
-    }
-
-    /**
-     * @return array
-     */
-    private function getMetadata()
-    {
-        $metadataFactory = new ClassMetadataFactory;
-        $metadataFactory->setEntityManager($this->getDoctrine()->getManager());
-        return $metadataFactory->getAllMetadata();
-    }
-
-    /**
-     * @return array
-     */
-    private function getClasses()
-    {
-        $classes = array();
-        /**
-         * @var ClassMetadata $class
-         */
-        foreach ($this->getMetadata() as $class) {
-            $classes[$class->getName()] = $class;
-        }
-        ksort($classes);
-        return $classes;
-    }
-
-    /**
-     * @param $classes
-     * @return string
-     */
-    private function generateGraph($classes)
-    {
-        $metagrapher = new MetadataGrapher;
-        $graph = $metagrapher->generateFromMetadata($classes);
-        return $graph;
-    }
-
-    /**
-     * @param string $dsl_text
-     * @return string
-     */
-    private function getGraphUrl($dsl_text){
-        $curl = new Curl(self::YUML_POST_URL);
-        $curl->setPosts(array('dsl_text' => $dsl_text));
-        $return = $curl->getResponse();
-
-        return self::YUML_REDIRECT_URL . $return;
+        return $this->redirect($yumlClient->getGraphUrl($yumlClient->makeDslText()));
     }
 }
