+++ lib/Onurb/Bundle/YumlBundle/Resources/config/services.yml
@@ -4,3 +4,7 @@ services:
         arguments: []
         tags:
             - { name: data_collector, template: OnurbYumlBundle:Toolbar:toolbar.html.twig, id: "doctrine_yuml" }
+    onurb.doctrine_yuml.client:
+        class: Onurb\Bundle\YumlBundle\Yuml\YumlClient
+        arguments:
+            - @doctrine.orm.entity_manager
